team-manager-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
sqlalchemy = "^2.0.48"
aiosqlite = "^0.22.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
openpyxl = "^3.1.5"
reportlab = "^4.4.10"
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
